Fix inconsistency between cavity decay and photon number operators.
Essentially the mapping between the different bases was defined for
annihilation operators for the initial models, but in terms of creation
operators for the mirror decay terms.  Everything is now defined in
terms of the mappings between creation operators in different
polarisation bases, and this has been updated in
v-Stirap-3vl-complex-biref and the mma notebooks.

All other vStirap-xxx notebooks remain to be updated.  This will be done
later after continuity checking between the new notebooks and the
already derived results.