Fix psi parsing between consecutive ham simulations
Off-diagonal elements in the calculation of the final density matrix of a single qutip.mesolve simulation were not calulated & stored correctly. Thus, the simulation of a subsequent hamiltonian with the final state from the previous simulation leads to an erroneous result.

Added arhoDict dictionary and used it to correctly store the final state in psiList after one iteration.